move recordToDataCite() to firebase functions to access admin credentials
