+++ firebase-functions/functions/index.js
@@ -1,7 +1,7 @@
 const admin = require("firebase-admin");
 const { translate } = require("./translate");
 const { checkURLActive } = require("./serverUtils");
-const { createDraftDoi, updateDraftDoi, deleteDraftDoi, getDoiStatus } = require("./datacite");
+const { createDraftDoi, updateDraftDoi, deleteDraftDoi, getDoiStatus, recordToDataCite } = require("./datacite");
 const { notifyReviewer, notifyUser } = require("./notify");
 const {
   updatesRecordCreate,
@@ -26,3 +26,4 @@ exports.deleteDraftDoi = deleteDraftDoi;
 exports.updateDraftDoi = updateDraftDoi;
 exports.getDoiStatus = getDoiStatus;
 exports.checkURLActive = checkURLActive;
+exports.recordToDataCite = recordToDataCite;
